# SILO-plateforme-source-2026-07-27/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 3C Studio API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
    post:
      operationId: addComment
      tags: [comments]
      summary: Add a comment to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
...
  /feature-flags/state:
    get:
      operationId: getFeatureFlagsState
      tags: [feature-flags]
      summary: Full feature flags state (admin/PM)
      responses:
        "200":
          description: Full state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureFlagsState"
        "403":
          description: Forbidden
  /feature-flags/global/{key}:
    put:
      operationId: setGlobalFeatureFlag
      tags: [feature-flags]
      summary: Set or reset a global flag (admin only)
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlobalFlagUpdate"
      responses:
        "200":
          description: Updated state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureFlagsState"
        "403":
          description: Forbidden
  /feature-flags/delegations/{key}:
    put:
      operationId: setFeatureFlagDelegation
      tags: [feature-flags]
      summary: Delegate (or revoke) a delegable key to the PM (admin only)
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DelegationUpdate"
      responses:
        "200":
          description: Updated state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureFlagsState"
        "403":
          description: Forbidden
  /feature-flags/overrides:
    post:
      operationId: createFeatureFlagOverride
      tags: [feature-flags]
      summary: Add a targeted override (admin, or PM if delegated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureFlagOverrideInput"
      responses:
        "201":
          description: Override created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureFlagOverride"
        "403":
          description: Forbidden
  /feature-flags/overrides/{id}:
    delete:
      operationId: deleteFeatureFlagOverride
      tags: [feature-flags]
      summary: Remove an override (admin, or PM if delegated)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Override removed
        "403":
          description: Forbidden
...
components:
  schemas:
...
    Comment:
      type: object
      required:
        [id, projectId, userId, authorName, content, type, isAdmin, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        userId:
          type: string
        authorName:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [comment, correction]
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [content, authorName]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
        type:
          type: string
          enum: [comment, correction]
        authorName:
          type: string
          minLength: 1
          maxLength: 200
...
    FeatureFlagsState:
      type: object
      required: [global, overrides, pmDelegations]
      properties:
        global:
          type: object
          additionalProperties:
            type: boolean
        overrides:
          type: array
          items:
            $ref: "#/components/schemas/FeatureFlagOverride"
        pmDelegations:
          type: array
          items:
            type: string
    FeatureFlagOverride:
      type: object
      required: [id, featureKey, scope, target, enabled, createdBy]
      properties:
        id:
          type: integer
        featureKey:
          type: string
        scope:
          type: string
          enum: [role, user, project, subscription]
        target:
          type: string
        enabled:
          type: boolean
        createdBy:
          type: string
          enum: [admin, pm]
    FeatureFlagOverrideInput:
      type: object
      required: [featureKey, scope, target, enabled]
      properties:
        featureKey:
          type: string
          minLength: 1
        scope:
          type: string
          enum: [role, user, project, subscription]
        target:
          type: string
        enabled:
          type: boolean
    GlobalFlagUpdate:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: ["boolean", "null"]
          description: null resets the flag to its default
    DelegationUpdate:
      type: object
      required: [delegated]
      properties:
        delegated:
          type: boolean